Changed to building the engine as a static library instead of as a dynamic library

+++ OpenGL-Engine/src/Engine/Input.h
@@ -11,8 +11,8 @@ namespace Engine {
 		inline static bool IsKeyPressed(int keyCode) { return s_Instance->IsKeyPressedImpl(keyCode); }
 		inline static bool IsMouseButtonPressed(int button) { return s_Instance->IsMouseButtonPressedImpl(button); }
 		inline static std::pair<float, float> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
-		inline static int GetMouseX() { return s_Instance->GetMouseXImpl(); }
-		inline static int GetMouseY() { return s_Instance->GetMouseYImpl(); }
+		inline static float GetMouseX() { return s_Instance->GetMouseXImpl(); }
+		inline static float GetMouseY() { return s_Instance->GetMouseYImpl(); }
 
 	protected:
 		// Implementations of input polling to be done in the platform-specific files
